requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.6.7
decorator==5.1.1
Django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
executing==1.2.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.1
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mysqlclient==2.1.1
nest-asyncio==1.5.6
openapi-codec==1.3.2
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.0.2
requests==2.28.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.1
traitlets==5.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
yarg==0.1.9
